fix(redact): the binary was never the reason, so every subprocess redacts its stderr
Closes #120.

The rule that a subprocess's stderr goes through the redactor arrived
narrow twice and was widened twice, and both corrections are the same
correction. It first qualified a function only if it called pushAuthEnv,
on the reasoning that a push is where a credential reaches git; that
missed EnsureHead's fetch and the merge-base ladder. Then it qualified on
the command being git, and named helm, kustomize and kubeconform as the
next question. This is that question, and the answer both times was that
the binary was never the reason.

What makes a child's stderr dangerous is that this process starts it
while holding credentials. cmd.Env is nil at nearly every call site, and
a nil Env means the child gets os.Environ() -- so a chart render runs
with GIT_TOKEN, ARGOCD_TOKEN and the model key in its environment, and so
does kubeconform. A plugin, a debug flag or a chart hook that prints its
environment puts them on stderr, and stderr is what these quoted into an
error that reaches a log and the gate's published report. The other half
is the one git already demonstrated: a chart render pulls from a registry
over somebody else's network, and a host that echoes a request header
back inside an error body is echoing a credential it was sent.

So three call sites redact -- gate/sources.go's runner, gate/validate.go's
kubeconform, agent/schemas.go's CRD read -- and the rule loses its second
condition rather than gaining a third. runsGit tested for exec.Command
*and* the literal "git", the second half being awkward precisely because
the binary is not always an argument to exec.Command: the push sites build
a []string of steps, so the name sits three statements from the call.
Dropping the binary from the rule dropped the awkwardness with it.
gitstderr_test.go is subprocess_stderr_test.go, because the old name had
stopped being true.

The self-check goes from eight to eleven, and the comment records both
moves it has made -- nine to eight when the clones were consolidated,
eight to eleven here -- because a floor lowered to make a blind walk pass
is indistinguishable from a floor lowered because a call site went away,
and only checking the walk against the code tells them apart. The three
added are the CRD read, the kubeconform run, and gate's own runner, which
is one function that is helm, `kustomize build` or `kubectl kustomize`
depending on which binary its caller handed it -- so counting it as a helm
site would not add up for anyone repeating the arithmetic.

A behavioural test beside the syntactic one, because a walk proving
redact.Text is called proves nothing about what it removes: a real
subprocess writing a primed secret to stderr, and the error that comes
back. Its sibling proves an unprimed process still reports what the tool
said, so the two cannot both pass for the wrong reason. Mutation-tested
in both directions -- blinding the walk fails on the floor, removing any
one redact.Text fails on that site by name.

Where the derivation stops is named, and there is a third escape now that
the binary condition is gone: a command run through .Output() with no
Stderr of its own has what it printed handed back on exec.ExitError.Stderr,
which this walk cannot see. internal/helmtest reads it that way twice.
That is test support -- both functions take a *testing.T, neither is in
the binary, and the redactor is never primed in a test -- so it is left
alone rather than carved out of the rule. Every .Output() in the shipped
code either sets its own Stderr, and so is counted, or discards the error.

The last acceptance criterion asked for a decision rather than a change:
whether these subprocesses should be given an explicit cmd.Env instead of
inheriting every credential this process loaded. They should, and not
here. That is a change to what helm runs with rather than to what bosun
prints; it wants a denylist derived from config.go's own envSecret calls,
because an allowlist that misses HELM_* or a proxy variable breaks a
chart render in a deployment nobody here can see. Filed as #122, with the
shape written down, and named in the safety model where the gap now sits.
