+++ CONTRIBUTING.md
@@ -40,7 +40,7 @@ on its own:
 | The ClusterRole covers the API reads the code makes | Not yet. A missing grant is a 403 that `cluster/` turns into a soft, honest sentence, so nothing would ever look | `chart_rbac_test.go` against `cluster.Reads()`, which the request paths are built from |
 | The prompt, the schema and the struct state one response shape | The hand-written list meant to hold two of them together omitted `escalationReason` | `llm/contract_test.go`, all four statements, every list derived |
 | Every credential `config.go` reads primes the process redactor | Not yet. A credential added to `Config` and wired to its one client compiles, passes, and has no symptom until a host echoes it back inside an error string | `redaction_test.go`, both sets derived: the credentials from `config.go`'s syntax tree, the coverage from a real `Config` of sentinels through `redact.Text`, and `main`'s own call from its syntax tree, so deleting the priming fails rather than passing |
-| Every git subprocess redacts what git wrote to stderr | Yes, and the first rule to catch it was itself too narrow: it qualified a function only if it called `pushAuthEnv`, which left the fetch, the ladder and three clones quoting git verbatim into a published report | `gitstderr_test.go`, deriving from the mechanism and not the call sites — a function that starts `git` and reads the buffer it gave that command, across every package rather than `gitprovider` alone |
+| Every subprocess redacts what it wrote to stderr | Yes, and the rule arrived too narrow twice: first it qualified a function only if it called `pushAuthEnv`, which left the fetch, the ladder and three clones quoting git verbatim into a published report; then it qualified on the command being git, which left `helm`, `kustomize` and `kubeconform` doing the same. The binary was never the reason | `subprocess_stderr_test.go`, deriving from the mechanism and not the call sites — a function that starts a subprocess and reads the buffer it gave it, across every package |
 | No credential this process holds reaches a command line | Yes, twice. The push spelled its token into the remote URL until it was moved into the environment, and a credential in `GIT_REPO_URL` went the same way on three separate clones after that | `gitremote_test.go`, deriving from git's own remote-facing subcommands that only `gitprovider` may run one; plus `pushremote_test.go` and `cloneargv_test.go`, which replace git with a shim and read the argv the kernel was actually given |
 | The shape of every MCP tool result, which somebody else's agent parses | Not yet, and this side of it cannot break loudly: the other half is a client in a repository this CI will never run. A renamed field, or an absence that becomes an empty array, is a silent break there and a passing build here | `mcp/testdata/*.json` through `mcp/contract_test.go` -- golden files, so a schema cannot drift without a reviewer seeing the diff, and `mcp_credentials_test.go`, which walks the result types from `mcp.Tools()` and the credential names from `config.go`'s syntax tree |
 | The gate's verdict, copied into `mcp`'s own shapes because `mcp` may not import `gate` | Not yet. A field the composition root forgets to copy is simply missing from every answer, and neither package has a compiler or a test that can see the other half | `mcp_gate_test.go` -- a real `DiffResult` through the real `Summarise` and the real adapter, with the blocker breakdown compared field by field off the struct rather than by name |
