+++ docs/safety-model.md
@@ -149,7 +149,7 @@ named them rather than a model.
 | Cannot compare a range it cannot establish | a chart version and the git tags of the project it packages frequently use different numbering. Refs come from the project's own release tags or from the publisher's recorded build revision; when neither meets the promotion's versions, no comparison is made and the note says why. Two refs picked out of the wrong numbering return real commits from a range that is not this one |
 | Cannot mutate the cluster | live reads are `get` and `list` only, and the chart's ClusterRole has no `create`, `update`, `patch` or `delete` verb anywhere. They are off by default: everything else the agent reads is public or already in the pull request, and this reads the operator's cluster |
 | Cannot send its own credentials to the model | the git token, the GitHub App private key, the ArgoCD token, the promotion token and the model API key are read once at start-up by `config.go` and handed to one client each. `prompt/` names none of them, and nothing that builds a prompt has a reference to reach. The API key is the only one that touches the provider at all, as the `Authorization` header of the call; it is not in the body. The model is sent the pull request, the gate's report and the rendered output, and it returns a verdict and a proposal |
-| Cannot publish a credential a host quoted back at it | every credential `config.go` loads primes one process-wide redactor at start-up -- plus one embedded in `GIT_REPO_URL`, which is not loaded as a credential but is handed to a remote as one -- and text this process did not author goes through it before it is logged, posted, or wrapped into an error. Today that is what `git` prints, on every command this process runs: both pushes, including the installation token minted for that one push which start-up never saw, and every clone, fetch, merge-base and diff, which talk to a host this process authenticated to. This is the second line and not the first: the first is that each credential is handed to one client and nothing else holds a reference to reach. It is also the narrow claim -- redaction removes this process's own secrets from a string, and does not make a string safe. It does not extend to `helm`, `kustomize` or `kubeconform`, whose stderr is quoted unredacted |
+| Cannot publish a credential a host quoted back at it | every credential `config.go` loads primes one process-wide redactor at start-up -- plus one embedded in `GIT_REPO_URL`, which is not loaded as a credential but is handed to a remote as one -- and text this process did not author goes through it before it is logged, posted, or wrapped into an error. Today that is what every subprocess prints: both pushes, including the installation token minted for that one push which start-up never saw; every clone, fetch, merge-base and diff, which talk to a host this process authenticated to; and the chart renders and schema checks, which talk to a registry and inherit every credential on the way. This is the second line and not the first: the first is that each credential is handed to one client and nothing else holds a reference to reach. It is also the narrow claim -- redaction removes this process's own secrets from a string, and does not make a string safe. The binary is not what makes a child's stderr dangerous; being started by something holding credentials is. What it is not is the first line: `cmd.Env` is nil at nearly every call site, so those children still inherit every credential this process loaded, and stopping that is #122 |
 | Cannot hand a credential to anything that reads a command line | a credential an operator embedded in `GIT_REPO_URL` never reaches git's argv, where `/proc/<pid>/cmdline` is world-readable. `gitprovider.Remote` splits that URL into the address git is given and an `http.<remote>.extraHeader` it is handed through the environment instead, which `/proc` exposes only to the owner. The remote stored as `origin` is clean, so the credential is attached per command, to the commands that contact a host and to no others -- ArgoCD's arrangement, for ArgoCD's reason. Structural rather than reviewed, in one direction: a derived test permits only `gitprovider` to run a git command that contacts a remote, and `agent`, `gateservice` and `supervisor` hold a `Remote` where each used to hold a URL string, so there is nothing in those packages left to pass. The configured string still exists in `config.go`, and `gitprovider.GitHub` still holds it to build a push remote from -- that path strips the userinfo itself, and has since before this. The same URL is stripped before it becomes the status page's repository link |
 | Cannot read a Secret | with `liveReads.enabled: false` the core API group is not granted at all, and with `liveReads.scope: groups` it is granted no further than `pods, events`. The chart creates no Role over Secrets anywhere. `liveReads.scope: wide` grants `apiGroups: ["*"]` cluster-wide and **can** read Secrets everywhere. RBAC has no deny rules and no way to subtract the core group, which is why "everything except Secrets" is not a setting |
 | Cannot be given a Secret read it does not need | the gate's cluster inventory comes from ArgoCD's API, not from the cluster Secrets those clusters are stored in. That read could not be made small enough: the gate wants four fields, and RBAC has no predicate for "the labels but not the data". There are no deny rules, `resourceNames` does not apply to `list`, and the apiserver applies the request's label selector *after* authorising. `GET /api/v1/clusters` serves the same four fields with the credential block redacted, which draws that line. The trade is real: an ArgoCD account token to mint and rotate, and a component that can be down on its own |
